pyproject.toml
[build-system]
# 64+ is broken with editable:
# - https://github.com/pypa/setuptools/issues/3518
# - https://github.com/python/mypy/issues/13392
requires = ['setuptools>=62.0, <64.0', 'wheel', 'setuptools_dso~=2.9']
build-backend = 'setuptools.build_meta'

[project]
name        = 'tilequant'
description = 'Tool for quantizing image colors using tile-based palette restrictions'
version     = '1.0.0'
readme      = 'README.rst'
requires-python = '>=3.8'
authors = [
  { name='Marco "Capypara" Köpcke', email='hello@capypara.de' },
]
classifiers = [
    'Development Status :: 4 - Beta',
    'Programming Language :: Python',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Topic :: Multimedia :: Graphics',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = [
    'Pillow~=9.5.0',
    'ordered-set~=4.1.0',
    'sortedcollections~=2.1.0',
    'click~=8.1.3',
    'setuptools_dso~=2.9'
]

[project.urls]
Source = 'https://github.com/SkyTemple/tilequant/'

[project.scripts]
tilequant = 'tilequant.cli:main'

[tool.mypy]
warn_unused_configs = true
mypy_path = '.'
explicit_package_bases = true
namespace_packages = true
check_untyped_defs = true
exclude = ['tilequant_dsoinfo.py', 'build/']

[[tool.mypy.overrides]]
module = 'setuptools_dso'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'setuptools_dso.probe'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'setuptools'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'sortedcollections'
ignore_missing_imports = true
